# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AckIn": {
        "properties": {
          "ids": {
            "items": {
              "type": "integer"
            },
            "title": "Ids",
            "type": "array"
          }
        },
        "title": "AckIn",
        "type": "object"
      },
...
      "BackupMeta": {
        "properties": {
          "bytes": {
            "description": "Stoerrelsen paa den lagrede bloben.",
            "title": "Bytes",
            "type": "integer"
          },
          "client_ts": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Klientens tidsstempel for oeyeblikksbildet. Godtas baade som epoke-millisekunder og ISO-8601 ved opplasting.",
            "title": "Client Ts"
          },
          "device_id": {
            "title": "Device Id",
            "type": "string"
          },
          "escrowed": {
            "description": "Om noekkelen er deponert - altsaa om kopien kan gjenopprettes UTEN gjenopprettingskoden. Lar en ny telefon vite det foer den laster ned 16 MB.",
            "title": "Escrowed",
            "type": "boolean"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "bytes",
          "schema_version",
          "device_id",
          "client_ts",
          "updated_at",
          "escrowed"
        ],
        "title": "BackupMeta",
        "type": "object"
      },
...
      "EscrowIn": {
        "properties": {
          "key_material": {
            "maxLength": 1024,
            "minLength": 4,
            "title": "Key Material",
            "type": "string"
          }
        },
        "required": [
          "key_material"
        ],
        "title": "EscrowIn",
        "type": "object"
      },
      "EscrowMateriale": {
        "properties": {
          "key_material": {
            "description": "Base64 av ugjennomsiktige byte. Serveren tolker dem ikke.",
            "title": "Key Material",
            "type": "string"
          },
          "updated_at": {
            "description": "Sist BRUKEREN deponerte. En intern omkryptering ved utskiftning av server-hemmeligheten flytter den ikke.",
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "key_material",
          "updated_at"
        ],
        "title": "EscrowMateriale",
        "type": "object"
      },
      "EscrowStatus": {
        "properties": {
          "escrowed": {
            "title": "Escrowed",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "escrowed",
          "updated_at"
        ],
        "title": "EscrowStatus",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Bestefar backend",
    "version": "0.2"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/backup": {
      "delete": {
        "description": "Sletter BLOBEN. En eventuell deponert noekkel blir staaende - den aapner\ningenting naar bloben er borte, og valget «gjenopprett uten kode» er en\ninnstilling brukeren har slaatt paa, ikke en foelge av at det finnes en\nkopi akkurat naa. Ville vi slettet noekkelen her, ville neste opplasting\nstilltiende vaert udeponert selv om bryteren stod paa.\nKontosletting (§9) fjerner begge deler.",
        "operationId": "delete_backup_v1_backup_delete",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Backup",
        "tags": [
          "backup"
        ]
      },
      "get": {
        "operationId": "download_backup_v1_backup_get",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Backup",
        "tags": [
          "backup"
        ]
      },
      "put": {
        "description": "Lagrer bloben. `client_ts` er klientens tidsstempel for oeyeblikksbildet.\n\nVERN MOT UTDATERT ENHET: er `client_ts` eldre enn den lagrede, svarer vi 409\ni stedet for aa overskrive. Spec §2 sier last-write-wins per post-ID, men\nden regelen kan bare haandheves KLIENT-side - serveren ser ikke inn i den\nkrypterte bloben. Uten dette vernet ville en gammel telefon som synker for\nfoerste gang paa maaneder kunne viske ut alt som er logget siden.\nKlienten kan overstyre med `force=true` naar brukeren har tatt et\nbevisst valg (f.eks. «gjenopprett fra denne enheten»).",
        "operationId": "upload_backup_v1_backup_put",
        "parameters": [
          {
            "in": "query",
            "name": "client_ts",
            "required": true,
            "schema": {
              "format": "date-time",
              "title": "Client Ts",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "schema_version",
            "required": false,
            "schema": {
              "default": 1,
              "title": "Schema Version",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "device_id",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 64,
              "title": "Device Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "force",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force",
              "type": "boolean"
            }
          },
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Backup",
        "tags": [
          "backup"
        ]
      }
    },
    "/v1/backup/key-escrow": {
      "delete": {
        "description": "Idempotent, og krever IKKE at hemmeligheten er konfigurert: aa slutte aa\ndeponere noekkelen skal alltid gaa gjennom. Den motsatte veien (503) ville\nlaast brukeren inne i et valg hen vil ut av.",
        "operationId": "slett_noekkel_v1_backup_key_escrow_delete",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Slett Noekkel",
        "tags": [
          "backup"
        ]
      },
      "get": {
        "operationId": "hent_noekkel_v1_backup_key_escrow_get",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowMateriale"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Hent Noekkel",
        "tags": [
          "backup"
        ]
      },
      "put": {
        "description": "Lagrer noekkelmaterialet til backup-bloben, kryptert i ro.\n\nMERK HVA DETTE BETYR: serveren har da baade bloben og noekkelen, og kan\nlese jaktloggen. Det er en bevisst avveining brukeren gjoer mot risikoen\nfor aa miste alt sammen med telefonen, og UI-teksten skal si det rett ut.\nIdempotent - en ny PUT erstatter materialet.",
        "operationId": "deponer_noekkel_v1_backup_key_escrow_put",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscrowIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EscrowStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deponer Noekkel",
        "tags": [
          "backup"
        ]
      }
    },
    "/v1/backup/meta": {
      "get": {
        "description": "Metadata uten selve bloben - slik at «har jeg en backup aa gjenopprette?»\npaa en ny telefon ikke krever nedlasting av hele nyttelasten.",
        "operationId": "backup_meta_v1_backup_meta_get",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupMeta"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Backup Meta",
        "tags": [
          "backup"
        ]
      }
    },
...
    "/v1/messages/ack": {
      "post": {
        "description": "Klienten kvitterer naar meldingen er VIST. Serveren sletter ikke raden -\nden markeres som levert, saa en klient som krasjer mellom henting og\nvisning ikke mister meldingen for godt.",
        "operationId": "ack_messages_v1_messages_ack_post",
        "parameters": [
          {
            "description": "«Bearer <access-token>» (§1).",
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "«Bearer <access-token>» (§1).",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AckIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ack Messages",
        "tags": [
          "meldinger"
        ]
      }
    },
...
  }
}